projects/hangul-writing/requirements.txt
flask==2.2.3
flask-cors==3.0.10
pillow==9.4.0
requests==2.28.2
werkzeug==2.2.3
gunicorn==20.1.0
numpy==1.24.2
protobuf==4.22.3
torch>=2.0.0  # Will be overridden by CUDA-specific install
torchvision>=0.15.0
transformers==4.37.2
accelerate==0.28.0
bitsandbytes==0.41.3
huggingface-hub>=0.16.4
flask-caching
safetensors>=0.4.0
nvidia-ml-py>=11.495.46
peft>=0.7.0
sentencepiece>=0.1.99
scipy>=1.11.0
